backend/requirements.txt
# AutoQA Backend Dependencies

# Server
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Screenshots & Image Processing
mss==9.0.1
Pillow==10.2.0

# Input Control
pyautogui==0.9.54

# Window Management (platform-specific)
pygetwindow==0.0.9; sys_platform == "win32"
pyobjc-framework-Quartz==10.1; sys_platform == "darwin"
pyobjc-framework-ApplicationServices==10.1; sys_platform == "darwin"

# HTTP Client (for Gemini API)
requests==2.31.0

# Document Processing
PyMuPDF==1.24.0
python-docx==1.1.0

# Video Processing
opencv-python==4.9.0.80

# Environment
python-dotenv==1.0.1
